:bug: next proxy 기본 타입아웃 설정 변경. 이제 600초의 timeout 시간을 가집니다.
